requirements.txt
numpy==2.2.1
pyopengl==3.1.7
opencv-python==4.10.0
pyrr==0.10.3
pypng==0.0.20
glfw==2.8.0
./deps/OpenGLContext-2.3.0-py2.py3-none-any.whl
PyVRML97==2.3.1
open3d==0.18.0
plyfile==1.1
tyro==0.9.6
rich==13.9.4